Added more PLU items, more fixes, finalization
More PLU items to the array, added documentation comments, made it so that when you get it wrong and get jumpscared, it clears input (it didnt before).

On rst.html, opening the pluspreadsheet opens in a new tab.

On pluSpreadsheet.html, fixed the table going upwards and being unable to scroll + added the rest of the PLU item names and codes